Library: rearrange launch into the base policy
this is a mere rearrangement of code (+lots of comments),
but helps to structure the overall construction better.

ThreadWrapper::launchThread() now does the actual work to build
the active std::thread object and assign it to the thread handle,
while buildLauncher is defined in the context of the constructors
and deals with wiring the functors and decaying/copying of arguments.